Move simple_path to its own file
